refactor: added validation schema
