+++ src/controllers/SelectYearController.ts
@@ -1,16 +1,16 @@
+import Joi from '@hapi/joi';
 import { SessionMiddleware } from 'ch-node-session-handler';
 import { PenaltyList } from 'ch-sdk-node/dist/services/lfp';
-import { Request } from 'express';
 import { controller } from 'inversify-express-utils';
 import { SafeNavigationBaseController } from './SafeNavigationBaseController';
-import { Validator } from './validators/Validator';
+import { FormValidator } from './validators/FormValidator';
 
 import { AuthMiddleware } from 'app/middleware/AuthMiddleware';
 import { PenaltyReferenceRouter } from 'app/middleware/PenaltyReferenceRouter';
 import { Appeal } from 'app/models/Appeal';
-import { PenaltyDetailsRadioComponent } from 'app/models/components/PenaltyDetailsRadioComponent';
+import { createPenaltyRadioButton } from 'app/models/components/PenaltyRadioButton';
+import { createSchema } from 'app/models/fields/PenaltyChoice.schema';
 import { PENALTY_DETAILS_PAGE_URI, REVIEW_PENALTY_PAGE_URI, SELECT_YEAR_PAGE_URI } from 'app/utils/Paths';
-import { ValidationResult } from 'app/utils/validation/ValidationResult';
 
 const template = 'select-the-year';
 
@@ -23,11 +23,7 @@ const navigation = {
     }
 };
 
-class RadioButtonComponentValidator implements Validator {
-    public async validate(_: Request): Promise<ValidationResult> {
-        return Promise.reject();
-    }
-}
+export const errorMessage: string = 'Select the penalty you want to appeal';
 
 // tslint:disable-next-line: max-classes-per-file
 @controller(SELECT_YEAR_PAGE_URI, SessionMiddleware, AuthMiddleware, PenaltyReferenceRouter)
@@ -36,7 +32,9 @@ export class SelectYearController extends SafeNavigationBaseController<any> {
     public static PENALTY_EXPECTED_ERROR: string = 'Penalty object expected but none found';
 
     constructor() {
-        super(template, navigation, new RadioButtonComponentValidator());
+        super(template, navigation, new FormValidator(Joi.object({
+            yearSelection: createSchema(errorMessage)
+        })));
     }
 
     public prepareViewModelFromAppeal(appeal: Appeal): Record<string, any> & any {
@@ -47,6 +45,6 @@ export class SelectYearController extends SafeNavigationBaseController<any> {
             throw new Error(SelectYearController.PENALTY_EXPECTED_ERROR);
         }
 
-        return { radioButtonComponent: PenaltyDetailsRadioComponent(penaltyList.items) };
+        return { penalties: penaltyList.items.map(createPenaltyRadioButton) };
     }
 }
